Refactor to make registering model configurations public
